% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_request}
\alias{.make_request}
\title{Função interna para fazer requisições HTTP}
\usage{
.make_request(url, params = NULL, timeout = 60)
}
\arguments{
\item{url}{URL para requisição}

\item{params}{Parâmetros da query string}

\item{timeout}{Timeout em segundos}
}
\value{
Lista com resposta JSON parseada
}
\description{
Função interna para fazer requisições HTTP
}
\keyword{internal}
