% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadados_filterable_columns}
\alias{get_metadados_filterable_columns}
\title{Retrieve list of filterable columns from dataset metadata}
\usage{
get_metadados_filterable_columns(dataset = "antt_pracaspedagio")
}
\arguments{
\item{dataset}{Name of the ONTL dataset (default: "antt_pracaspedagio")}
}
\value{
Tibble with two columns: dataset and column_name
}
\description{
Generates a data frame with columns that can be used to filter
data in the API, facilitating specific query construction.
}
\examples{
\dontrun{
columns <- get_metadados_filterable_columns()
print(columns)
}
}
